Oh lots of changes
Now downloads and detects Mac OS for java. Can find the IP of the
machine now. Lots of different interface tweaks and additions of
functions to streamline things a bit.

Also got the whitelist feature working! The script now generates a
white-list.txt file that corresponds with the server. That one took a
lot of learning, turns out this is done a lot better with the split
method rather than using re to detect inputs.